Remove unused agenda job
