+++ packages/back-end/src/init/queue.ts
@@ -1,6 +1,5 @@
 import addExperimentResultsJob from "back-end/src/jobs/updateExperimentResults";
 import refreshFactTableColumns from "back-end/src/jobs/refreshFactTableColumns";
-import refreshFactTableTopValues from "back-end/src/jobs/refreshFactTableTopValues";
 import updateScheduledFeatures from "back-end/src/jobs/updateScheduledFeatures";
 import addWebhooksJob from "back-end/src/jobs/webhooks";
 import addMetricUpdateJob from "back-end/src/jobs/updateMetrics";
@@ -36,7 +35,6 @@ export async function queueInit() {
   updateStaleInformationSchemaTable(agenda);
   expireOldQueries(agenda);
   refreshFactTableColumns(agenda);
-  refreshFactTableTopValues(agenda);
   addSdkWebhooksJob(agenda);
   updateLicenseJob(agenda);
