+++ packages/back-end/src/jobs/refreshFactTableColumns.ts
@@ -17,7 +17,6 @@ import { DataSourceInterface } from "back-end/types/datasource";
 import { getContextForAgendaJobByOrgId } from "back-end/src/services/organizations";
 import { trackJob } from "back-end/src/services/otel";
 import { logger } from "back-end/src/util/logger";
-import { runColumnTopValuesQuery } from "back-end/src/jobs/refreshFactTableTopValues";
 
 const JOB_NAME = "refreshFactTableColumns";
 type RefreshFactTableColumnsJob = Job<{
@@ -59,6 +58,36 @@ const refreshFactTableColumns = trackJob(
     await updateFactTableColumns(factTable, updates);
   }
 );
+
+export async function runColumnTopValuesQuery(
+  context: ReqContext,
+  datasource: DataSourceInterface,
+  factTable: Pick<FactTableInterface, "sql" | "eventName">,
+  column: ColumnInterface
+): Promise<string[]> {
+  if (!context.permissions.canRunFactQueries(datasource)) {
+    context.permissions.throwPermissionError();
+  }
+
+  const integration = getSourceIntegrationObject(context, datasource, true);
+
+  if (
+    !integration.getColumnTopValuesQuery ||
+    !integration.runColumnTopValuesQuery
+  ) {
+    throw new Error("Top values not supported on this data source");
+  }
+
+  const sql = integration.getColumnTopValuesQuery({
+    factTable,
+    column,
+    limit: 100,
+  });
+
+  const result = await integration.runColumnTopValuesQuery(sql);
+
+  return result.rows.map((r) => r.value);
+}
 
 export async function runRefreshColumnsQuery(
   context: ReqContext,
